+++ grammar/ChangeLog.md
@@ -1,5 +1,9 @@
 This file logs the changes in FST files, starting from the DSDSmor fork.
 
+2023-11-27:
+
+* add inflection class `Frac` for fractions
+
 2023-11-24:
 
 * add inflection classes `IPro-unsereiner` and `IPro-unsereins` for indefinite